+++ packages/webcomponents/src/components/payment-form/payment-form.tsx
@@ -15,9 +15,10 @@ import JustifiAnalytics from '../../api/Analytics';
 })
 export class PaymentForm {
   @Prop() bankAccount?: boolean;
-  @Prop() card?: boolean = true;
+  @Prop() card?: boolean;
   @Prop() email?: string;
-  @Prop() clientId: string;
+  @Prop() clientId?: string;
+  @Prop() authToken?: string;
   @Prop() accountId?: string;
   @Prop() submitButtonText?: string;
 
@@ -35,6 +36,14 @@ export class PaymentForm {
 
   componentWillLoad() {
     this.analytics = new JustifiAnalytics(this);
+    if (!this.validateProps()) {
+      this.errorEvent.emit({
+        errorCode: ComponentErrorCodes.MISSING_PROPS,
+        message: 'clientId or authToken is required',
+        severity: ComponentErrorSeverity.ERROR
+      });
+      this.submitButtonEnabled = false;
+    }
   }
 
   connectedCallback() {
@@ -60,9 +69,33 @@ export class PaymentForm {
     this.submitButtonEnabled = false;
   }
 
+  showPaymentMethodTypeSelector() {
+    return this.card && this.bankAccount;
+  }
+
+  getSelectedPaymentMethodType() {
+    if (this.showPaymentMethodTypeSelector()) {
+      return this.selectedPaymentMethodType;
+    } else if (this.card) {
+      return PaymentMethodTypes.card;
+    } else if (this.bankAccount) {
+      return PaymentMethodTypes.bankAccount;
+    } else {
+      return PaymentMethodTypes.card;
+    }
+  }
+
   paymentMethodSelectedHandler(event: CustomEvent) {
     const paymentMethodType: PaymentMethodTypes = event.detail;
     this.selectedPaymentMethodType = paymentMethodType;
+  }
+
+  private validateProps(): boolean {
+    return !!(this.clientId || this.authToken);
+  }
+
+  private getToken(): string {
+    return this.authToken || this.clientId;
   }
 
   async submit(event) {
@@ -79,7 +112,7 @@ export class PaymentForm {
     try {
       const billingFormFieldValues = await this.billingFormRef.getValues();
       const paymentMethodData = { email: this.email, ...billingFormFieldValues };
-      const tokenizeResponse = await this.paymentMethodFormRef.tokenize(this.clientId, paymentMethodData, this.accountId);
+      const tokenizeResponse = await this.paymentMethodFormRef.tokenize(this.getToken(), paymentMethodData, this.accountId);
       if (tokenizeResponse.error) {
         this.errorEvent.emit({
           errorCode: ComponentErrorCodes.TOKENIZE_ERROR,
@@ -103,17 +136,17 @@ export class PaymentForm {
     return (
       <Host>
         <form class="row gy-3">
-          {this.card && this.bankAccount && (
+          {this.showPaymentMethodTypeSelector() && (
             <div class="col-12">
               <justifi-payment-method-selector
-                selectedPaymentMethodType={this.selectedPaymentMethodType}
+                selectedPaymentMethodType={this.getSelectedPaymentMethodType()}
                 onPaymentMethodSelected={event => this.paymentMethodSelectedHandler(event)}
               />
             </div>
           )}
           <div class="col-12">
             <justifi-payment-method-form
-              payment-method-form-type={this.selectedPaymentMethodType}
+              payment-method-form-type={this.getSelectedPaymentMethodType()}
               iframeOrigin={config.iframeOrigin}
               ref={el => {
                 if (el) {