+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @repo/docs
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [5820727]
+- Updated dependencies [4f834b7]
+- Updated dependencies [f304a5e]
+- Updated dependencies [d9b1e8c]
+  - @justifi/webcomponents@4.15.0
+
+## 0.1.0
+
+### Minor Changes
+
+- 0183890: Add `authToken` prop to PaymentForm
+  Update `Authorization` section in docs
+
+### Patch Changes
+
+- Updated dependencies [0189492]
+- Updated dependencies [0183890]
+  - @justifi/webcomponents@4.14.0
+
 ## 0.0.10
 
 ### Patch Changes